requirements.txt
# Streamlit and related visualization tools
streamlit==1.39.0
streamlit-chat==0.1.1
matplotlib==3.9.2
seaborn==0.13.2

## microservices
fastapi[standard]
fastapi-users[beanie]
uvicorn
pydantic-settings

# AI and OpenAI integration
openai==1.52.0
tiktoken==0.8.0
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.8

# Azure SDKs
azure-identity==1.19.0
azure-mgmt-resource==23.1.1
# azure-ai-documentintelligence==1.0.0b3
azure-ai-documentintelligence==1.0.2

azure-cosmos==4.7.0
azure-search-documents==11.6.0b5
azure-storage-blob==12.23.1
azure-monitor-opentelemetry==1.6.4

# General utilities
requests==2.32.3
python-dotenv==1.0.1
pandas==2.2.3
pydantic==2.9.2
pyarrow==17.0.0
aiohttp==3.10.10
backoff==2.2.1
PyPDF2==3.0.1
Jinja2==3.1.4
pymongo==4.10.1
colorama
PyMuPDF
rapidfuzz

# HTTPX fix for OpenAI
httpx==0.27.2

# Async and event loop tools
asyncio==3.4.3
tenacity==8.5.0

semantic_kernel
azure-monitor-opentelemetry
azure-ai-inference[opentelemetry]
opentelemetry-exporter-otlp
azure-ai-projects
opentelemetry-instrumentation-httpx
